delete only works if beverage has no servings

+++ app/Http/Controllers/BeveragesController.php
@@ -27,7 +27,11 @@ class BeveragesController extends Controller
     }
 
     public function destroy(Beverage $beverage) {
-        $beverage->delete();
+        if ($beverage->servings->count()) {
+            // soft delete
+        } else {
+            $beverage->delete();
+        }
         return redirect('/beverages');
     }
